+++ src/main/fc/stats.h
@@ -5,7 +5,7 @@
 typedef struct statsConfig_s {
     uint32_t stats_total_time; // [Seconds]
     uint32_t stats_total_dist; // [Metres]
-    uint32_t stats_flight_count;
+    uint16_t stats_flight_count;
 #ifdef USE_ADC
     uint32_t stats_total_energy; // deciWatt hour (x0.1Wh)
 #endif
@@ -19,7 +19,7 @@ void statsOnDisarm(void);
 
 #else
 
-#define statsInit() do {} while (0)
+#define statsInit()     do {} while (0)
 #define statsOnArm()    do {} while (0)
 #define statsOnDisarm() do {} while (0)
